# docs/swagger.yaml
basePath: /
definitions:
  models.CreateSubscriptionRequest:
    properties:
      end_date:
        example: 12-2024
        type: string
      price:
        example: 299
        minimum: 0
        type: integer
      service_name:
        example: Netflix
        type: string
      start_date:
        example: 01-2024
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  models.SubscriptionResponse:
    properties:
      end_date:
        example: 12-2024
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      price:
        example: 299
        type: integer
      service_name:
        example: Netflix
        type: string
      start_date:
        example: 01-2024
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
  models.TotalCostResponse:
    properties:
      total_cost:
        example: 1499
        type: integer
    type: object
  models.UpdateSubscriptionRequest:
    properties:
      end_date:
        example: 12-2024
        type: string
      price:
        example: 599
        type: integer
      service_name:
        example: Netflix Premium
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A service for managing user subscriptions
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Subscription Aggregator API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: List subscriptions from the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SubscriptionResponse'
            type: array
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription for a user
      parameters:
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SubscriptionResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Subscription already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Delete a subscription by ID
      parameters:
      - description: Subscription ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No content
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Subscription not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a subscription
      tags:
      - subscriptions
    get:
      description: Get a single subscription by its ID
      parameters:
      - description: Subscription ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SubscriptionResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Subscription not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a subscription by ID
      tags:
      - subscriptions
    patch:
      consumes:
      - application/json
      description: Update an existing subscription
      parameters:
      - description: Subscription ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Updated subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SubscriptionResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Subscription not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a subscription
      tags:
      - subscriptions
  /subscriptions/total-cost:
    get:
      description: Calculate total cost of subscriptions with optional filters
      parameters:
      - description: User ID
        format: uuid
        in: query
        name: user_id
        type: string
      - description: Service name (partial match)
        in: query
        name: service_name
        type: string
      - description: Start date filter
        format: MM-YYYY
        in: query
        name: start_date
        type: string
      - description: End date filter
        format: MM-YYYY
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TotalCostResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get total cost of subscriptions
      tags:
      - subscriptions
schemes:
- http
- https
swagger: "2.0"